+++ frontend/src/index.js
@@ -11,6 +11,7 @@ import relativeTime from "dayjs/plugin/relativeTime";
 import red from "@material-ui/core/colors/deepOrange";
 import blue from "@material-ui/core/colors/indigo";
 import grey from "@material-ui/core/colors/grey";
+import amber from "@material-ui/core/colors/amber";
 
 import Main from "./pages/Main/Main";
 import LoginPageContainer from "./pages/Login/LoginPageContainer";
@@ -30,11 +31,17 @@ const store = configureStore(history);
 
 const muiTheme = createMuiTheme({
   palette: {
-    primary: blue,
+    primary: {
+      main: blue[500]
+    },
     secondary: red,
+    error: red,
+    warning: amber,
+    info: blue,
     grey: {
       main: grey[100]
-    }
+    },
+    tonalOffset: 0.6
   },
   typography: {
     useNextVariants: true